requirements.txt
pyyaml
requests
flask